refactor: small improvements (#620)
* process_input_note: simplified advice stack params names
* kernel: using stdlib utility to extract digest
* prologue: added details about nullifier
* tests: explain why Send/Sync is being enforced
* removed contraints on the struct definition

The type constraints should be added into the `impl`s blocks, because
the constraints on the `struct`s propagate up and makes composition very
verbose.

This also renamed the mod data to data_store to better represent its
content.
